functions/python/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
google-genai==1.0.0
gunicorn==23.0.0
Pillow==10.4.0
gevent==24.2.1
qdrant-client
google-cloud-storage